backend/requirements.txt
fastapi
uvicorn
browser-use
playwright
langchain-ollama
langchain
python-dotenv
websockets
pytest
pytest-asyncio
httpx
